Fixed edge case for creator and contributor too
Problem was that if a pre-filled form was deleted and an empty form
took its index, Django would compare the empty form's values to
the filled form's initial values. Inside the validation mechanism,
it would look like the empty form had been changed. That causes
Django to validate it, see that fields are not filled in, and
send it back to the user to fix.

My workaround is to override `is_valid()` calls and first check
if all of the fields in a form are their respective empty values.
If so, the form's data is made an empty dict and it automatically
returns as valid. My existing code to skip empty forms in the JSON
dicts omits them then.