Bibtex form user experience complete
Now set up so that the user can regenerate the text citation from the bibtex,
and must do so or deliberately unlock the submission button. This should
avoid accidentally changing the bibtex without updating the text citation,
but also allow users to make minor edits to the bibtex without changing the
text.